move processor-lifecycle-manager from messaging to service module
