+++ sdk/spring/spring-cloud-azure-autoconfigure/src/test/java/com/azure/spring/cloud/autoconfigure/servicebus/AzureServiceBusProcessorClientConfigurationTests.java
@@ -10,9 +10,10 @@ import com.azure.spring.cloud.autoconfigure.TestBuilderCustomizer;
 import com.azure.spring.cloud.service.implementation.servicebus.factory.ServiceBusSessionProcessorClientBuilderFactory;
 import com.azure.spring.cloud.service.servicebus.consumer.ServiceBusErrorHandler;
 import com.azure.spring.cloud.service.servicebus.consumer.ServiceBusRecordMessageListener;
-import com.azure.spring.messaging.servicebus.implementation.core.ServiceBusProcessorClientLifecycleManager;
+import com.azure.spring.cloud.service.implementation.servicebus.lifecycle.ServiceBusProcessorClientLifecycleManager;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.autoconfigure.AutoConfigurations;
+import org.springframework.boot.test.context.FilteredClassLoader;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
 
 import static com.azure.spring.cloud.autoconfigure.servicebus.ServiceBusTestUtils.CONNECTION_STRING_FORMAT;
@@ -273,7 +274,32 @@ class AzureServiceBusProcessorClientConfigurationTests {
                 assertThat(context).hasSingleBean(AzureServiceBusProcessorClientConfiguration.class);
                 assertThat(context).hasSingleBean(ServiceBusProcessorClient.class);
                 assertThat(context).doesNotHaveBean(ServiceBusProcessorClientLifecycleManager.class);
+
+                ServiceBusProcessorClient processorClient = context.getBean(ServiceBusProcessorClient.class);
+                assertThat(processorClient.isRunning()).isFalse();
             });
+    }
+
+    @Test
+    void withoutClassShouldNotConfigureLifecycleManager() {
+        contextRunner
+                .withClassLoader(new FilteredClassLoader(ServiceBusProcessorClientLifecycleManager.class))
+                .withPropertyValues(
+                        "spring.cloud.azure.servicebus.processor.entity-name=test-queue",
+                        "spring.cloud.azure.servicebus.processor.entity-type=queue",
+                        "spring.cloud.azure.servicebus.processor.connection-string=" + String.format(CONNECTION_STRING_FORMAT, "test-namespace")
+                )
+                .withUserConfiguration(AzureServiceBusPropertiesTestConfiguration.class)
+                .withBean(ServiceBusRecordMessageListener.class, () -> messageContext -> { })
+                .withBean(ServiceBusErrorHandler.class, () -> errorContext -> { })
+                .run(context -> {
+                    assertThat(context).hasSingleBean(AzureServiceBusProcessorClientConfiguration.class);
+                    assertThat(context).hasSingleBean(ServiceBusProcessorClient.class);
+                    assertThat(context).doesNotHaveBean(ServiceBusProcessorClientLifecycleManager.class);
+
+                    ServiceBusProcessorClient processorClient = context.getBean(ServiceBusProcessorClient.class);
+                    assertThat(processorClient.isRunning()).isFalse();
+                });
     }
 
     @Test
