+++ sdk/spring/spring-cloud-azure-service/src/test/java/com/azure/spring/cloud/service/implementation/servicebus/lifecycle/ServiceBusProcessorClientLifecycleManagerTest.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-package com.azure.spring.messaging.servicebus.implementation.core;
+package com.azure.spring.cloud.service.implementation.servicebus.lifecycle;
 
 import com.azure.messaging.servicebus.ServiceBusProcessorClient;
 import org.junit.jupiter.api.Test;
